Fix telegram bot

+++ src/api/user/router.py
@@ -20,16 +20,17 @@ async def edit_profile(*, current_user: AuthenticatedUser = Depends(get_current_
                        edit_info: EditProfileRequest = Body(...)) -> OkResponse:
     user = await USER.get_by_id(current_user.id)
 
-    if user.avatar_id is None:
-        avatar_id = str(uuid.uuid4())
-        await USER.edit_profile(current_user.id, avatar_id=avatar_id)
-    else:
-        await remove_file(user.avatar_id)
-        avatar_id = user.avatar_id
+    if avatar:
+        if user.avatar_id is None:
+            avatar_id = str(uuid.uuid4())
+            await USER.edit_profile(current_user.id, avatar_id=avatar_id)
+        else:
+            await remove_file(user.avatar_id)
+            avatar_id = user.avatar_id
 
-    await async_upload_file(file=avatar,
-                            file_id=avatar_id,
-                            ext='.jpg')
+        await async_upload_file(file=avatar,
+                                file_id=avatar_id,
+                                ext='.jpg')
 
     result = await USER.edit_profile(current_user.id, **edit_info.dict(exclude_none=True))
     if not result:
